Refactor settings helper (#98)
* Move classes to correct folder

The declared namespace and the location of the class files did not
match, which can result in a class not found error.

* Extract user placeholder validation

Moved user placeholder validation logic to a separate function to
decouple it from ACL owner validation. This allows for user placeholder
validation to be reused in contexts that do not require the other
checks, improving code modularity and reusability.